+++ .gradle/buildOutputCleanup/cache.properties
@@ -1,0 +1,2 @@
+#Mon Nov 14 17:52:28 WEST 2022
+gradle.version=7.3.1
