+++ src/main/resources/META-INF/MANIFEST.MF
@@ -1,3 +1,3 @@
 Manifest-Version: 1.0
-Main-Class: pihole.WidgetApplication
+Main-Class: controllers.WidgetApplication
